+++ mangadex_downloader/format/base.py
@@ -36,6 +36,7 @@ class BaseFormat:
         compress_img,
         replace,
         no_verify,
+        no_chapter_info,
         kwargs_iter_chapter_img
     ):
         self.path = path
@@ -43,6 +44,7 @@ class BaseFormat:
         self.compress_img = compress_img
         self.replace = replace
         self.verify = not no_verify
+        self.no_chapter_info = no_chapter_info
         self.kwargs_iter = kwargs_iter_chapter_img
 
     def get_images(self, chap_class, images, path, count):