+++ mangadex_downloader/format/pdf.py
@@ -393,10 +393,10 @@ class PDFSingle(PDF):
             img_name = count.get() + '.png'
             img_path = path / img_name
 
-            get_chapter_info(chap_class, img_path, self.replace)
-            images.append(img_path)
-
-            count.increase()
+            if not self.no_chapter_info:
+                get_chapter_info(chap_class, img_path, self.replace)
+                images.append(img_path)
+                count.increase()
 
             images.extend(self.get_images(chap_class, chap_images, path, count))
 
@@ -462,9 +462,10 @@ class PDFVolume(PDF):
                 img_name = count.get() + '.png'
                 img_path = volume_path / img_name
 
-                get_chapter_info(chap_class, img_path, self.replace)
-                images.append(img_path)
-                count.increase()
+                if not self.no_chapter_info:
+                    get_chapter_info(chap_class, img_path, self.replace)
+                    images.append(img_path)
+                    count.increase()
 
                 images.extend(self.get_images(chap_class, chap_images, volume_path, count))